requirements.txt
PySimpleGUI
psgtray
pyinstaller
pyyaml